requirements.txt
# core
pandas >= 1.5

# environment variable loader
python-dotenv >= 1.0

# JSON schema validation
jsonschema >= 4.0

#Claude Agent SDK
claude-agent-sdk >= 0.1

# typing extras for python 3.10 compatibility
typing-extensions > 4.0

tabulate>=0.9 